refactor(autonomi): move data/archive/fs modules
