+++ autonomi/src/client/vault/user_data.rs
@@ -8,10 +8,10 @@
 
 use std::collections::HashMap;
 
-use crate::client::archive::ArchiveAddr;
-use crate::client::archive_private::PrivateArchiveAccess;
 use crate::client::data::GetError;
 use crate::client::data::PutError;
+use crate::client::files::archive::ArchiveAddr;
+use crate::client::files::archive_private::PrivateArchiveAccess;
 use crate::client::payment::PaymentOption;
 use crate::client::registers::RegisterAddress;
 use crate::client::vault::VaultError;
